Prepare CTRL for C24_WMDE_Mobile_DE_07
Based on the VAR of test 06

Ticket: https://phabricator.wikimedia.org/T378706
