+++ banners/mobile/C24_WMDE_Mobile_DE_07/styles/styles.scss
@@ -1,6 +1,7 @@
 @use 'src/themes/Mikings/swatches/skin_default' with (
 	$upgrade-to-yearly-button-form: true,
 	$progress-bar: true,
+	$soft-close: true,
 	$thank-you-box: true
 );
 @use 'src/components/BannerConductor/banner-transition';
@@ -24,4 +25,5 @@
 	$pagination-padding: 10px,
 	$pagination-height: 40px
 );
+@use 'src/themes/Mikings/SoftClose/SoftClose';
 @use 'src/themes/Mikings/ProgressBar/ProgressBar';
