--- banners/mobile/C24_WMDE_Mobile_DE_07/styles/styles_var.scss
@@ -1,29 +1,0 @@
-@use 'src/themes/Mikings/swatches/skin_default' with (
-	$upgrade-to-yearly-button-form: true,
-	$progress-bar: true,
-	$soft-close: true,
-	$thank-you-box: true
-);
-@use 'src/components/BannerConductor/banner-transition';
-@use 'src/themes/UseOfFunds/swatches/skin_default' as uof-default;
-@use 'src/themes/UseOfFunds/UseOfFunds';
-@use 'src/themes/Mikings/defaults';
-@use './Banner';
-@use './MiniBanner' with (
-	$height: 336px
-);
-@use './FullPageBanner';
-@use 'src/themes/Mikings/Footer/Footer';
-@use 'src/themes/Mikings/ThankYouBox/ThankYouBox';
-@use 'src/themes/Mikings/Footer/SelectionInput';
-@use 'src/themes/Mikings/DonationForm/MultiStepDonation';
-@use 'src/themes/Mikings/DonationForm/Forms/UpgradeToYearlyButtonForm';
-@use 'src/themes/Mikings/DonationForm/SubComponents/SelectGroup';
-@use 'src/themes/Mikings/DonationForm/SubComponents/SelectCustomAmount';
-@use 'src/themes/Mikings/DonationForm/SubComponents/SmsBox';
-@use 'src/themes/Mikings/Slider/Slider' with (
-	$pagination-padding: 10px,
-	$pagination-height: 40px
-);
-@use 'src/themes/Mikings/SoftClose/SoftClose';
-@use 'src/themes/Mikings/ProgressBar/ProgressBar';
